Feature/account screen

+++ app/src/main/java/com/ort/waynimovil/ui/components/TransactionList.kt
@@ -1,11 +1,9 @@
 package com.ort.waynimovil.ui.components
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -13,7 +11,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
@@ -24,7 +21,7 @@ data class TransactionItemData(
     val date: String,
     val description: String,
     val aut: String,
-    val amount: Float
+    val amount: Double
 )
 
 @Composable
@@ -71,18 +68,18 @@ fun TransactionListPreview() {
     WaynimovilTheme {
         TransactionList(
             items = listOf(
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", +500.0f),
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0f),
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0f),
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0f),
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", +500.0f),
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0f),
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0f),
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0f),
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", +500.0f),
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0f),
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0f),
-                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", +500.0f),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", +500.0),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", +500.0),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", +500.0),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", -500.0),
+                TransactionItemData("12/12/2021", "Compra en supermercado", "Aut. 394991", +500.0),
             )
         )
     }
